backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "live-transcription-demo-backend"
version = "0.1.0"
description = "Backend for the live transcription coaching demo"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115,<1",
  "httpx>=0.28,<1",
  "numpy>=1.26,<3",
  "pydantic>=2.9,<3",
  "pydantic-settings>=2.6,<3",
  "pyyaml>=6,<7",
  "sounddevice>=0.5,<1",
  "uvicorn>=0.32,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "pytest-asyncio>=0.24,<1",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
